refactor(assistant): remove legacy chat, fold slide-out shell into the assistant feature under components/
