+++ app/features/assistant/components/index.ts
@@ -1,0 +1,7 @@
+// Full-page workspace (also reused inside the slide-out panel)
+export { AssistantWorkspace } from './assistant-workspace';
+
+// Global slide-out shell: provider + navbar trigger + panel
+export { AssistantProvider, useAssistant } from './assistant-context';
+export { AssistantPanel } from './assistant-panel';
+export { AssistantTrigger } from './assistant-trigger';
