+++ app/features/assistant/hooks/use-chat-logic.ts
