+++ app/routes/dashboard/index.tsx
@@ -1,5 +1,5 @@
 import type { Route } from './+types/index';
-import { AssistantWorkspace, STAFF_ASSISTANT_CONFIG } from '@/features/assistant-workspace';
+import { AssistantWorkspace, STAFF_ASSISTANT_CONFIG } from '@/features/assistant';
 import { HEADER_STACK_H } from '@/features/milo/lib/dimensions';
 import { useEnv } from '@/hooks';
 import { useApp } from '@/providers/app.provider';
